+++ Cargo.toml
@@ -10,10 +10,7 @@ repository = "https://github.com/johanhelsing/noisy_bevy"
 version = "0.13.0"
 
 [dependencies]
-bevy = { version = "0.18", default-features = false, features = [
-    "bevy_asset",
-    "bevy_shader",
-] }
+bevy = { version = "0.18", default-features = false }
 
 [dev-dependencies]
 bevy = { version = "0.18", default-features = false, features = [
@@ -28,3 +25,17 @@ bevy_egui = { version = "0.39", default-features = false, features = [
     "default_fonts",
 ] }
 insta = "1.21"
+
+[features]
+default = [
+    "gpu",
+    "cpu",
+]
+
+gpu = [
+    "bevy/bevy_asset",
+    "bevy/bevy_shader",
+]
+
+cpu = []
+
